创建分支优化

+++ ADDP-INF/build.sh
@@ -18,5 +18,8 @@ cp ${jar} ./ADDP-INF/${APP}.jar
 #        docker rm ${APP}:${ENV}
 #    fi
 #}
-docker rmi ${APP}:${ENV}
-docker build -t ${APP}:${ENV}   --build-arg APP_NAME=${APP} --build-arg ENV=${ENV} -f ./ADDP-INF/Dockerfile ./ADDP-INF
+if [ "$ENV" != "bak" ]
+    then
+        docker rmi ${APP}:${ENV}
+        docker build -t ${APP}:${ENV}   --build-arg APP_NAME=${APP} --build-arg ENV=${ENV} -f ./ADDP-INF/Dockerfile ./ADDP-INF
+fi
